+++ src/rules/index.js
@@ -3,7 +3,7 @@ import atFunctionPattern from "./at-function-pattern"
 import atMixinPattern from "./at-mixin-pattern"
 import dollarVariableNoMissingInterpolation from "./dollar-variable-no-missing-interpolation"
 import dollarVariablePattern from "./dollar-variable-pattern"
-import selectorNoRedundantParent from "./selector-no-redundant-parent"
+import selectorNoRedundantNestingSelector from "./selector-no-redundant-nesting-selector"
 
 export default {
   "at-extend-no-missing-placeholder": atExtendNoMissingPlaceholder,
@@ -11,5 +11,5 @@ export default {
   "at-mixin-pattern": atMixinPattern,
   "dollar-variable-no-missing-interpolation": dollarVariableNoMissingInterpolation,
   "dollar-variable-pattern": dollarVariablePattern,
-  "selector-no-redundant-parent": selectorNoRedundantParent,
+  "selector-no-redundant-nesting-selector": selectorNoRedundantNestingSelector,
 }